🚚 Move project functions to a subdirectory
