+++ src/functions/project/get-artefact-destination-cloud-run.ts
@@ -3,7 +3,7 @@ import {
   ProjectGetArtefactDestination,
   ServiceContainerConfiguration,
 } from '@causa/workspace-core';
-import { GoogleConfiguration } from '../configurations/index.js';
+import { GoogleConfiguration } from '../../configurations/index.js';
 
 /**
  * Implements the {@link ProjectGetArtefactDestination} function for Cloud Run services.
